+++ planner/core/casetest/partition/testdata/integration_partition_suite_in.json
